Add integration tests for the Elasticsearch checkPlatform method.
